Move materialized view tests to separate class
